+++ sdkclient/client_test.go
@@ -1,4 +1,4 @@
-package tx
+package sdkclient
 
 import (
 	"testing"
